// types.ts
export interface ListeningQuestion {
  question: string;
  options: {
    key: string;
    value: string;
  }[];
  correctAnswer: string;
}

export interface ListeningTask {
  script: string;
  questions: ListeningQuestion[];
}